`janus_cli`: add DAP message decoder
`janus_cli decode-dap-message` can read binary encoded DAP wire messages
from either stdin or a file and then decode them as a specified message
type, printing the `std::fmt::Debug` representation of the decoded
message to stdout.

`janus_cli` uses some of the scaffolding from the `janus` server
binaries for option parsing, and so it used to expect to be provided a
configuration file with logging and particularly datbase config. That
isn't needed for the `decode-dap-message` command, so we lower the
`CommonBinaryOptions` from `janus_cli::Options` down into those variants
of `janus_cli::Command` that actually need those flags.

Resolves #441