Ensure we check container readiness on IPv4
`interop_binaries::test_util::await_http_server` blocks until HTTP
requests are responded to on a specified port. The assumption is that
this port is bound on localhost with IPv4, but because we were
connecting to `localhost:{port}`, it could sometimes connect to the port
bound on IPv6, which might be bound by some totally different container,
meaning that tests prematurely will consider a container's port to be
available.

We now connect to `http://127.0.0.1:{port}`, which ought to ensure we
connect to IPv4.