Extract logs from integration test containers
Adds plumbing for extracting logs from Janus and Daphne containers on
test failures. We introduce trait `CopyLogs` and implement it on both.
For Janus, we recursively copy out the entire `/logs` directory from a
container. For Daphne, we run `docker logs` against the container.

Extracted logs are written to a directory on the host. If the
`JANUS_E2E_LOGS_PATH` environment variable is set, its value is used
(this will be used by CI so the log files can be uploaded as GitHub
Actions artifacts). If the env var is unset, a tempdir is created and
its path is logged.