Plumb `Prio3Aes128CountVec` VDAF throughout Janus
Adds support to Janus for an extra VDAF. Also plumbs it into the interop
test API, which required allowing the representation of vectors of
numbers to a `Measurement`, and unconditionally representing numbers as
strings to avoid JSON number precision problems. A change to
draft-dcook-ppm-dap-interop-test-design will follow to update the
language describing how to encode measurements and aggregate results.

Resolves #478