Store leader input shares unencrypted
When handling uploads, the leader now decrypts and decodes its input
share and the public share before recording them in the datastore. This
allows us to discard immediately reports that can never be aggregated
and simplifies the implementation of the aggregation job creator, since
it can now fetch well-formed values from the datastore.

See #442 for rationale.

Resolves #442