De-flake aggregator::tests::upload_wrong_hpke_config_id.
Previously, the HPKE config ID in the report was a fixed value (101).
Since one of the two HPKE config IDs configured for the task was a
random value, in 1 in 256 runs the random value would turn out to match
the fixed value. This would cause the HPKE config ID in the report to
not actually be wrong, which caused the test to fail.

Now, we choose a value to use in the report that is not used in the HPKE
configs of the task.