Introduce `CollectJobState::Deleted`
For #344, we introduce a "Deleted" state for a collect job. We have to
distinguish from the "Abandoned" state so that we can eventually return
the right HTTP status if a collector polls the collect job URI.

This commit defines the new state in the SQL schema and adds support for
entering it using `Datastore::update_collect_job`. Along the way, we
resolve #242: `put_collect_job` and `update_collect_job` now take
`&CollectJob` as an argument instead of just the field values to set.

Part of #344
Resolves #242