Correct errors for DAP-02 collect, agg share reqs
DAP-02 updates the "Batch Validation" section [1] with several new error
types. This commit updates `janus_server` with the new problem document
types, returns them as appropriate, and adds tests to verify that
correct errors are constructed.

The exception is `batchOverlap`: while the `collect_job_driver` does
correctly check for that condition when stepping a collect job, DAP-02
requires that this error be presented to the collector while handling a
`CollectReq`. Meeting that requirement will require more significant
refactoring that will arrive in a later pull request.

Part of #470

[1]: https://www.ietf.org/archive/id/draft-ietf-ppm-dap-02.html#section-4.5.6