`collect_job_driver`: check job state at finish (#580)
We now check whether the state of a collect job has changed while the
lease is held. This is possible if a DELETE request is sent to the
collect job URI while the collect job driver is working. We also check
for unexpected state transitions. Along the way, we expand the metrics
recorded by `collect_job_driver`.

Part of #344