Fixup Janus/Janus integration test.
Broken by a previous commit in this PR, the Janus/Janus integration test
set the leader & helper up to use differing VDAF verification keys. This
didn't work out so well. :-)

As part of the fix, allow TaskBuilders to be cloned; this makes it a lot
easier to generate paired tasks for leader & helper with all of the
relevant fields matching, without requiring the caller to generate those
values themselves.