Unify janus_{aggregator,core}::VdafInstance. (#693)
This cleanup has been possible since we started using "test-util"
features to conditionally-compile test-only behavior. I trip over this a
lot, and importing either type led to some awkward code, so I think it
is worthy of a quick cleanup.

Note that the serialization format of VdafInstance changed, which may
affect task serialization in our DB. I think this is OK since there is
no expectation of backwards-compatibility between 0.1 & 0.2.