Detect & skip build scripts for rust-analyzer.
rust-analyzer has a setting which disables build scripts. Unfortunately,
at least on my machine, this setting is ineffectual and build scripts
are run anyway. This implies the `cargo check` done by rust-analyzer was
actually building Janus. If the container build failed, the entire build
process would be aborted & VSCode would not show any error highlighting.

This new check is based on RUSTC_WRAPPER, which rust-analyzer sets by
default to the full path of the rust-analyzer binary. This environment
variable can be unset -- please don't do so. :-)

As a side-effect, I think this speeds up how quickly rust-analyzer will
be able to check Janus.