Detect & skip build scripts for rust-analyzer. (#707)
rust-analyzer has a setting which disables build scripts. Unfortunately, at least on my machine, this setting is ineffectual and build scripts are run anyway. This implies the `cargo check` done by rust-analyzer was actually building Janus as part of the container-build process. If the container build failed, the entire build process would be aborted & VSCode would not show any error highlighting in some cases.

This new check is based on `RUSTC_WRAPPER`, which rust-analyzer sets by default to the full path of the rust-analyzer binary. rust-analyzer can be configured to not set this environment variable -- please don't do so. :-)

As a side-effect, I think this speeds up how quickly rust-analyzer will be able to check Janus. Incremental container builds are quick, but not instantaneous.