DAP-03: implement current-batch functionality in Aggregator.
The logic for deciding which batch to return for current-batch is
currently very simple: we let Postgres decide which of the eligible
batches to return. We may well need to tweak this approach later (e.g.
to always return the least-recently-returned eligible batch), but this
should be fine for a first pass.