DAP-03: Update HTTP problem documents (#854)
* DAP-03: Update HTTP problem documents

DAP-03 has [new language][1] on HTTP problem documents. The fields
`taskid` and `instance` are no longer mandatory, so we don't have to
include the former when it's not known and we don't have to always set
the latter to `".."`.

[1]: https://datatracker.ietf.org/doc/html/draft-ietf-ppm-dap-03#name-errors

Resolves #545

* Remove `detail` field from HTTP problem documents

DAP-03 no longer requires us to populate this field in RFC 7807 problem
documents, and the value we were putting in there was identical to the
`title`, so we can just stop doing that for now.