replace `generate_vdaf_values` with `run_vdaf`
The tests in mod `datastore` defined a mechanism for running a VDAF and
recording a transcript, which is redundant given `run_vdaf` and
`VdafTranscript` in `janus_core`. This commit moves `datastore` tests to
`run_vdaf`. Additionally, `VdafTranscript` now has convenience methods
for getting preparation state and output shares, which absolves its
clients of writing gnarly `assert_matches!` macros that get so long that
`rustfmt` gives up on tidying them up.