`integration_tests`: don't panic in panic handler (#938)
The `Drop` implementation on `integration_tests::janus::Janus` tries to
copy logs out of interop test containers using `docker cp`. It would
panic if that failed, which would result in a double panic (because test
failures in `cargo test` are panics), and that would cause output from
failing tests to be lost, making it hard to diagnose failures.

With this change, we no longer panic if either we can't `exec(2)`
`docker` or if `docker cp` fails. Instead, the errors get printed to
stdout so at least they will appear in GitHub Actions logs, as will some
information about the test that failed.

Example test failure _without_ this change:
https://github.com/divviup/janus/actions/runs/4010819099/jobs/6887739020

Example test failure _with_ this change:
https://github.com/divviup/janus/actions/runs/4011174836/jobs/6888472737