Introduce client_reports.aggregation_started flag. (#988)
This new flag is used to track which client reports have been attached
to an aggregation, allowing us to create aggregation jobs without
needing to consult the report_aggregations table. Hopefully, this will
reduce the amount of contention between aggregation job creation & other
datastore operations in the system.

Note that we set the flag for both VDAFs without an aggregation
parameter, and VDAFs with an aggregation parameter. However, it's only
effectively used by VDAFs without an aggregation parameter. Still, the
flag is set for consistency.