Support write-batching for report uploads.
This introduces a couple of new configuration options on the aggregator
binary, allowing configuration of the batching behavior. Batching can be
disabled entirely by setting the batch size to 1 or less, or by reducing
the write delay to 0.

A few smaller optimizations while I was touching the report-upload code:
* Remove report read to check existence, in favor of updating the write
  to return a specific error code if the row already exists & driving
  the existence check based on that.
* For fixed-size tasks, remove the conflicting-collect-job check. Due to
  the way reports are allocated to batches, there is no possibility of
  collision here.

(These optimizations reduce the number of DB queries for all query
types: from 3 to 2 for time-interval tasks, and from 3 to 1 for
fixed-size tasks. For fixed-size tasks, these optimizations additionally
reduce the number of serialized network round-trips from 2 to 1!)

Technical note:
* I wrapped aggregator::Error in an Arc to work around the inability to
  clone errors in general. Another option would be to change Error to
  wrap an Arc<InnerError>, with InnerError being similar to the existing
  Error -- this would be a larger change, but would allow Error to
  implement Clone even if the inner error can't be cloned. For now, I
  leave this for later optimization (we might want to make the same
  change to all errors if we go this route).