Fix report replay check in aggregation job init
`handle_aggregate_init_generic` was checking if any of the report shares
in an `AggregationJobInitReq` already occurred in the `client_reports`
table, and would refuse to begin preparing an input share if so,
producing a `report_replayed` error.

However, this is not what DAP requires ([1]):

> Check if the report has already been aggregated *with this aggregation
> parameter*. If this check fails, the input share MUST be marked as invalid
> with the error `report_replayed`. This is the case if the report was used in
> a previous aggregate request and is therefore a replay.

We now check whether there's a `report_aggregations` row for the
`ReportId` and whether its corresponding aggregation job uses the same
aggregation parameter as the incoming aggregation job. Additionally, we
allow `put_report_share` to write the same report twice, because that's
perfectly OK for a VDAF like Poplar1.

[1]: https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/blob/89e5dc32be05ebd4eb032c7f6d0aaa7074c69b08/draft-ietf-ppm-dap.md?plain=1#L1331