Switch to Relaxed memory ordering for retry flag. (#1024)
This is fine because all we care about is total modification order,
which is still guaranteed by Relaxed memory ordering. (On x86, relaxed
atomic loads/stores are implemented as ordinary loads/stores with no
fencing required.)

Also, play a little code golf with the transaction code -- it was
getting a little hairy.