Add tests for collection job idempotence (#1036)
Adds new tests to verify that PUT and POST to collection jobs has the
expected idempotence semantics:

 - PUT to a collection job is idempotent
 - POST to a collection job allows the leader to update what "current
   batch" means in a fixed size task

As it turns out, our existing implementation already satisfied these
requirements, so this only adds test coverage. New tests and support
code are introduced in a new `collection_job_tests` module, to avoid
unnecessarily growing `aggregator/src/aggregator.rs`.

Existing tests on collection jobs defined in that file are renamed to
all start with `collection_job`, which makes it easier to run just those
tests that pertain to collection jobs. We could (should) move those
tests to the new module, but that would make this PR unnecessarily noisy
and would obscure the real changes it introduces.

Part of #998