Shard batch aggregations. (#1037)
Batch aggregations (which collect the aggregate share for each given
batch) are now sharded: writes during the aggregation process are sent
to a (randomly-chosen) shard; reads during the collection process read &
merge all shards to retrieve the final aggregate share.

The purpose of this change is to reduce write contention during the
aggregation process. The tradeoff is in increased read volume &
computation during the collection process.

The number of shards is controlled by a pair of new
`batch_aggregation_shard_count` configuration values. The aggregator
configuration value controls the helper sharding value, and the
aggregation job driver configuration value controls the leader sharding
value. (As this sharding behavior is private to the aggregator, there is
no reason that these values need to match up.)