ignore test for mutated replayed request for now
Per code review, we are going to check that replayed leader aggregation
continue requests haven't changed. That will be implemented in a
follow-up PR. For now, we flip the logic of the existing test so that it
expects a mutated replay to yield HTTP 400 and `unrecognizedMessage`,
but `#[ignore]` the test since that check is not implemented yet.