Replayed `AggregationJobContinueReq`s can't change (#1091)
* Add last request hash to aggregation jobs table

Add a column to `aggregation_jobs` to store the 32 byte SHA-256 hash of
the most recent `AggregationJobContinueReq` message, and affordances to
`datastore::models::AggregationJob` to work with it.

* Replayed `AggregationJobContinueReq` can't change

When we successfully handle an `AggregationJobContinueReq`, we now
record the SHA-256 hash of the request body. Then, if we later receive
another request trying to advance an aggregation job to the round it's
already on, we check the hash of the incoming request against the
recorded one. If they match, then we replay the previously-computed
prepare shares. If they don't, the request fails with HTTP 409 Conflict.

Resolves #994