Janus Aggregator API implementation. (#1113)
Based on Trillium; I was pretty happy with how the library worked out.

Note that we currently return the entire task -- including secrets --
wherever a task is returned (e.g. post task, get task). We had
previously discussed that the secrets would be returned by a different
endpoint. I think this can be changed later, and more than that, I think
the most common case will be adding a new task -- where the control
plane will need the secrets in order to relay them to the co-aggregator.