Use no secrets in `ci-build`
The `ci-build` workflow is not runnable from forks. This is because it
wants to authenticate to our private Google Artifact Repository in order
to fetch Daphne container images. However, those tests are disabled, and
going forward we'll test Daphne using the generic interop test interface
and using containers that team publishes to DockerHub. So we should have
no need to authenticate to private repositories and thus use secrets.
This commit removes mention of `DAPHNE_INTEROP_CONTAINER` from
`ci-build.yml` and deletes the steps that authenticate to GCP and the
Docker repository it hosts. This should make it possible for PRs on
forks to run CI.