Publish SQL migrations to minor version directory
Publish SQL migration scripts to a directory in the bucket named for the
minor version (e.g., "0.2") as well as the patch version (e.g.,
"0.2.17"). Going forward, downstream projects will consume migrations
from the minor version directory, but we publish to both places for now,
to ease the transition.

Part of #1240