Explicitly store last prepare steps for replay.
I did this because it turned out to be (basically) necessary for both
the ping-pong implementation, as well as the race-condition fixes.

This also fixes some buggy behavior in the corners of replay. For
example, if a round which ended up dropping a report was replayed, the
Helper would respond differently on replay: no prepare step at all the
first time, a failed prepare step with error ReportDropped on replay.
[There may be more; I didn't do a close reading of every code path.]