Run `datastore` tests against multiple schemas
We want to prove in tests that some Janus version can run safely on
multiple schema versions to make database schema migrations safer. In
this commit:

 - `aggregator_core::test_util::EphemeralDatastore` can now be
   constructed with a `max_schema_version` argument to control which
   migration scripts are applied during tests
 - `datastore::SUPPORTED_SCHEMA_VERSIONS` is now emitted by the
   `supported_schema_versions` macro, which also constructs a template
   for stamping out tests.
 - We adopt [`rstest`][1] to inject an `EphemeralDatastore` instance
   into tests in `aggregator_core::datastore::Datatore::tests`. Using
   [`rstest_reuse`][2], we can automatically stamp out versions of
   existing tests that run using multiple schema versions.

We only use this dependency injection technique in the `datastore`
module, because that should be the only module that's tightly coupled to
the database schema. We could run all tests that use a datastore this
way, at the cost of increasing overall test runtime.

[1]: https://docs.rs/rstest
[2]: https://docs.rs/rstest_reuse