Honor `DAP-Auth-Token` or `Authorization: Bearer`
When acting as an HTTP server (i.e., helpers handling aggregation job or
aggregate share messages, or leaders handling collection requests),
Janus will now check for the authentication token in the more standard
`Authorization: Bearer <token>` format defined in RFC 6750, as it
already does in the aggregator API. If that header is absent, it falls
back to checking `DAP-Auth-Token` for compatibility with existing peers.
Nothing changes about the format or handling of the authentication token
itself, which remains an opaque blob o' bytes.

Part of #472