Use opaque counter, rather than timestamp, to order SQL migrations. (#1304)
Using a timestamp to order SQL migrations can allow migrations to be
merged out-of-order with the order than SQLx expects to be able to apply
them. If a "later" migration was released before an "earlier" migration
was merged, our DB would be un-migratable and we would have a production
incident on our hands.