Fix flake in aggregation_job_acquire_release. (#1332)
Due to Postgres locking semantics, it is possible that concurrent
attempts to lease aggregation jobs may leave a job unleased that could
have been returned to one of the leasers. This is rare, but led to a
test flake when it occurred (it took about ~10m of looped runs of the
relevant test to trigger on my workstation). The fix is to retry
concurrent requests for some time, stopping once we get the expected
number of leases.

I think this doesn't mask a problem worth solving for production uses of
this code:
 * It is very rare for this behavior to be triggered.
 * The downside is that a given aggregation job's processing will be
   delayed; however, it is very likely to be picked up by the next
   attempt to lease.
 * I believe this can only happen when there are few aggregation jobs,
   i.e. when we have little/no backlog. If there are many aggregation
   jobs, all concurrent leasers should receive a full complement of
   aggregation jobs.