Do not reread batches which were read/written already.
The last thing the aggregation job writer does is, for any newly-closed
batches, find all related collection jobs and move them to state
COLLECTABLE if they are ready to do so. A collection job is ready to
move to state COLLECTABLE if all related batches are in state CLOSED.

This change reduces the amount we have to read. For fixed-size tasks,
this will always remove all of the batch reads, since every collection
job will be related to exactly one batch. For time-interval tasks, this
will reduce the amount we have to read, but may not eliminate it
entirely: collection jobs can be related to arbitrarily-long intervals
of batches which may not have been related to any of the aggregation
jobs being updated.