Remove INVALID report aggregation state.
This state was only used by the Leader, and only if the Helper sent an
unexpected preparation message (a Continue when a Finished was expected,
or vice-versa). It was treated by all program logic as equivalent to the
Failed state. This change replaces all uses of Invalid with
Failed(VdafPrepError), and eliminates the Invalid state entirely.