Add retry to initial batch writes.
Among other things, this will allow transactions to be retried if e.g.
the first two aggregation jobs for a given batch complete concurrently;
before this change, we would retry at the level of one of the
aggregation jobs, but this might be too late -- if the Helper did not
support retries of aggregation job initializations, one of the
aggregation jobs would be lost.

(I do not add a similar retry for aggregation job/report aggregation
writes: these entities are expected to be unique when generated, by way
of including a freshly-generated aggregation job ID. Batches, on the
other hand, are primarily identified by preexisting task ID, batch ID,
and aggregation parameters, and thus need to account for initial-write
conflicts.)