aggregation_job_writer: close batches only when all reports aggregating.
This semantics change should remove any need for integration test code
to wait or sleep for aggregation to settle, as long as the client report
uploads have a happens-before relationship with the collection process
starting.

This change might also be useful for a "real" collector: it provides the
guarantee that any client reports which have been received before
collection starts will go through the aggregation process before
collection is considered complete. So a real collector can issue a
collection request after the end of a given collection interval and
expect all reports received in that interval to be included in the
eventual collection response, no matter how behind the aggregations
might be.

Note that this check is racy: a batch can close _concurrently_ with a
report being uploaded to the batch. This is handled as before: the
report will eventually be marked BatchCollected by the aggregation
process & will not be reflected in the eventual collection result.