Datastore: implement barrier if we see "in failed tx" error.
For some error modes, Postgres will return an error to the caller & then
fail all future statements within the same transaction with an "in
failed SQL transaction" error. This effectively means one statement will
receive a "root cause" error and then all later statements will receive
an "in failed SQL transaction" error. In a pipelined scenario, if our
code is processing the results of these statements concurrently--e.g.
because they are part of a `try_join!`/`try_join_all` group--we might
receive & handle one of the "in failed SQL transaction" errors before we
handle the "root cause" error, which might cause the "root cause"
error's future to be cancelled before we evaluate it. If the "root
cause" error would trigger a retry, this would mean we would skip a
DB-based retry when one was warranted.

To fix this problem, we (internally) wrap all direct DB operations in
`run_op`. This function groups concurrent database operations into
"operation groups", which allow us to wait for all operations in the
group to complete (this waiting operation is called "draining"). If we
ever observe an "in failed SQL transaction" error, we drain the
operation group before returning.  Under the assumption that the "root
cause" error is concurrent with the "in failed SQL transactions" errors,
this guarantees we will evaluate the "root cause" error for retry before
any errors make their way out of the transaction code.