Remove CONCURRENTLY and IF EXISTS parameters on index deletion.
CONCURRENTLY is not allowed inside of a transaction. I think it's fair game just
to run them non-concurrently. If we're running this migration, it's very likely
that we're in a local development scenario where the database is mostly empty
and this deletion will happen quickly.

Removal of IF EXISTS was discussed in https://github.com/divviup/janus/issues/1311#issuecomment-1528163091--it's guarding against
a scenario that we _want_ to error out on, especially during testing.