Break dependency of report_aggregations on client_reports.
This is required to allow client reports to be GC'ed independently of
aggregation jobs; notably, this is the only foreign-key relationship
crossing the reports/aggregations/collections boundary.

The changes are, roughly:
  * The report_aggregations table now stores the client report ID,
    rather than a reference to a client_reports row, as well as the
    client_timestamp of the report.
  * At the start of the aggregation process, if a client report can't be
    found, it fails aggregation of that report (with a ReportDropped
    error) rather than failing the entire aggregtion.

Rather than copying input share information at time of the first
aggregation step, we could instead copy it at time of aggregation job
creation. The upside is that this would remove the possibility of the
client report being GC'ed between aggregation job creation & the first
step. However, I think aggregating reports that are actively being GC'ed
is a marginal enough case that it does not warrant the extra storage
complexity to make this happen.